bin: add a script to download openapi json
This script is used to download the openapi json from the server and
save it to a file.

In the next commit this file is then used by the client to generate
the client code.

This script is meant to be invoked from the root of the repository,
i.e. `./bin/save-openapi-json`.