Implement a flake.nix that uses the upstream llama.cpp flake by reference
This permits llama-cpp-python to be used with the upstream llama.cpp flake.nix deciding how to go about performing the build on the current platform, rather than rebuilding a local copy of libllama with separately-managed build instructions.

To use this, run:

```
nix develop
```

and one will be dropped into a shell in which

```
$ python -c 'import llama_cpp'
```

functions correctly, _without_ using cmake to build the `vendor/llama.cpp` source tree (instead, `LLAMA_CPP_LIB` is set to a `/nix/store` path containing `libllama.dylib` or `libllama.so` built according to the upstream llama.cpp `flake.nix` file's instructions).

As a future enhancement, a check should probably be added to ensure that the `flake.lock` pin and the `.gitmodules` dependency stay in sync.